requirements.txt
streamlit
pdfplumber
langchain
langchain-huggingface
langchain-groq
langchain-community
faiss-cpu
python-dotenv
PyPDF2
sentence-transformers
